More signedness cleanup

+++ pool.hpp
@@ -4,9 +4,9 @@
 struct stringpool {
 	unsigned long left = 0;
 	unsigned long right = 0;
-	unsigned long off = 0;
+	size_t off = 0;
 };
 
-long addpool(struct memfile *poolfile, struct memfile *treefile, const char *s, char type);
+size_t addpool(struct memfile *poolfile, struct memfile *treefile, const char *s, char type);
 
 #endif
